Clean up metamorphic test with helpers
